Introduce csv-to-json.
csv-to-json can convert a CSV stream into JSON stream.

The column names of the CSV stream are interpreted as a dot-separate path
to output object.

If --base-object-key is specified, the output object is initialized from the JSON encoded
object found in the specified column.

If a column is a valid JSON encoding of a JSON number, object, boolean or array
the value is encoded as that object, otherwise it is encoded as a JSON string.

Use --strings to force all values to be encoded as JSON strings.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>