Introduce json-to-csv.
json-to-csv allows mapping a JSON stream to a CSV stream.

The --columns parameter is used to specify both a number of paths into
an input JSON object and the name of the CSV column that the object
will be mapped to.

Each column of the output stream is either the JSON string found at the
specified object path (if a JSON string exists at that path) or the
JSON encoding of the object found at that path, otherwise.

If --base-object-key is specified, the JSON encoding of any part of the input object that is
not mapped by a CSV column is copied in the specified column

Signed-off-by: Jon Seymour <jon@wildducktheories.com>