Refactor the sort process into a number of composable types.
SortKeys can be converted into a Sortable, a sort.Interface
or a SortProcess.

Sortable is a sort.Interface that can be converted into a SortProcess.

Sortable.Comparators can be edited to add arbitrary comparators which
can be built from string comparators by calling Sortable.Comparator.

SortProcess can apply a Sort to a stream of CSV records that must have a specified Key.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>†